[core] Move the data buffer out of the ctap_response struct for more flexibility
For the CTAPHID layer, it is useful to have a contiguous memory buffer to hold both the CTAP status code (1 bytes) and the CTAP response (CBOR-encoded data) together.
Then the CTAPHID layer, can directly take that contiguous buffer as use it as the CTAPHID_CBOR response message payload.
We chose this particular solution to avoid introducing unnecessary dependency between CTAP and CTAPHID (we try to keep both layers completely independent).