[stm32h533] Fix app_hw_crypto_sha256_update() (HAL_HASH_Accumulate() length must be a multiple of 4)
HAL_HASH_Accumulate() requires all parts to be of a length that is a multiple of 4. Only the last part that must be processed via HAL_HASH_AccumulateLast() can have an arbitrary length. However, our hash_alg_t API enforce no such constraints. Therefore, we implement a 4-bytes buffer to handle this.